remove grpc.go and http.go
